split traces CLI by audience into module tree
