+++ crates/ironclaw_reborn_cli/src/commands/traces/mod.rs
@@ -29,6 +29,13 @@ use ironclaw_reborn_traces::contribution::{
     write_trace_policy_for_scope,
 };
 
+mod admin;
+mod contributor;
+mod reviewer;
+mod shared;
+mod tenant;
+mod worker;
+
 /// `traces` subcommand surface for the standalone Reborn CLI.
 #[derive(Debug, Args)]
 pub(crate) struct TracesCommand {
@@ -1965,813 +1972,57 @@ impl std::fmt::Display for TraceRetentionJobItemStatusArg {
 
 async fn run_traces(cmd: TracesSubcommand) -> anyhow::Result<()> {
     match cmd {
-        TracesSubcommand::OptIn {
-            endpoint,
-            user_scope,
-            bearer_token_env,
-            upload_token_issuer_url,
-            upload_token_issuer_allowed_hosts,
-            upload_token_audience,
-            upload_token_tenant_id,
-            upload_token_workload_token_env,
-            upload_token_invite_code,
-            upload_token_issuer_timeout_ms,
-            include_message_text,
-            include_tool_payloads,
-            scope,
-            selected_tools,
-            allow_pii_review_bypass,
-            min_submission_score,
-        } => opt_in(OptInOptions {
-            endpoint,
-            user_scope,
-            bearer_token_env,
-            upload_token_issuer_url,
-            upload_token_issuer_allowed_hosts,
-            upload_token_audience,
-            upload_token_tenant_id,
-            upload_token_workload_token_env,
-            upload_token_invite_code,
-            upload_token_issuer_timeout_ms,
-            include_message_text,
-            include_tool_payloads,
-            scope,
-            selected_tools,
-            allow_pii_review_bypass,
-            min_submission_score,
-        }),
-        TracesSubcommand::OptOut { user_scope } => opt_out(user_scope.as_deref()),
-        TracesSubcommand::Status { json, user_scope } => {
-            show_policy_status(json, user_scope.as_deref())
-        }
-        TracesSubcommand::Preview {
-            recorded_trace,
-            include_message_text,
-            include_tool_payloads,
-            scope,
-            channel,
-            engine_version,
-            contributor_id,
-            credit_account_ref,
-            output,
-            enqueue,
-        } => {
-            preview_recorded_trace(PreviewOptions {
-                recorded_trace,
-                include_message_text,
-                include_tool_payloads,
-                scope,
-                channel,
-                engine_version,
-                contributor_id,
-                credit_account_ref,
-                output,
-                enqueue,
-            })
-            .await
-        }
-        TracesSubcommand::Enqueue { envelope } => {
-            let envelope = load_envelope(&envelope)?;
-            let policy = read_policy()?;
-            enqueue_envelope_with_policy(
-                &envelope,
-                &policy,
-                TraceContributionAcceptance::QueueFromPreview,
-            )?;
-            println!(
-                "Queued redacted trace contribution {}",
-                envelope.submission_id
-            );
-            Ok(())
-        }
-        TracesSubcommand::FlushQueue { limit } => flush_queue(limit).await,
-        TracesSubcommand::QueueStatus { json, scope } => show_queue_status(json, scope.as_deref()),
-        TracesSubcommand::Credit {
-            json,
-            notice,
-            notice_scope,
-            ack,
-            snooze_hours,
-        } => show_credit(json, notice, notice_scope.as_deref(), ack, snooze_hours).await,
-        TracesSubcommand::Submit {
-            envelope,
-            endpoint,
-            bearer_token_env,
-        } => submit_envelope(&envelope, &endpoint, &bearer_token_env).await,
-        TracesSubcommand::ListSubmissions { json, summary } => list_submissions(json, summary).await,
-        TracesSubcommand::Revoke {
-            submission_id,
-            endpoint,
-            bearer_token_env,
-        } => revoke_submission(submission_id, endpoint.as_deref(), &bearer_token_env).await,
-        TracesSubcommand::IngestHealth { endpoint, json } => {
-            trace_commons_ingest_health(&endpoint, json).await
-        }
-        TracesSubcommand::QuarantineList {
-            endpoint,
-            lease_filter,
-            bearer_token_env,
-            json,
-        } => trace_commons_quarantine_list(&endpoint, &bearer_token_env, lease_filter, json).await,
-        TracesSubcommand::ActiveLearningReviewQueue {
-            endpoint,
-            limit,
-            privacy_risk,
-            lease_filter,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_active_learning_review_queue(
-                &endpoint,
-                &bearer_token_env,
-                limit,
-                privacy_risk,
-                lease_filter,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::ReviewDecision {
-            endpoint,
-            submission_id,
-            decision,
-            reason,
-            credit_points_pending,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_review_decision(TraceCommonsReviewDecisionOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                submission_id,
-                decision,
-                reason,
-                credit_points_pending,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ReviewLeaseClaim {
-            endpoint,
-            submission_id,
-            lease_ttl_seconds,
-            review_due_at,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_review_lease_claim(TraceCommonsReviewLeaseClaimOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                submission_id,
-                lease_ttl_seconds,
-                review_due_at,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ReviewLeaseClaimNext {
-            endpoint,
-            lease_ttl_seconds,
-            review_due_at,
-            privacy_risk,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_review_lease_claim_next(TraceCommonsReviewLeaseClaimNextOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                lease_ttl_seconds,
-                review_due_at,
-                privacy_risk,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ReviewLeaseClaimBatch {
-            endpoint,
-            limit,
-            lease_ttl_seconds,
-            review_due_at,
-            privacy_risk,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_review_lease_claim_batch(TraceCommonsReviewLeaseClaimBatchOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                limit,
-                lease_ttl_seconds,
-                review_due_at,
-                privacy_risk,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ReviewLeaseRelease {
-            endpoint,
-            submission_id,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_review_lease_release(&endpoint, &bearer_token_env, submission_id, json)
-                .await
-        }
-        TracesSubcommand::AppendCreditEvent {
-            endpoint,
-            submission_id,
-            event_type,
-            credit_points_delta,
-            reason,
-            external_ref,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_append_credit_event(TraceCommonsAppendCreditEventOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                submission_id,
-                event_type,
-                credit_points_delta,
-                reason,
-                external_ref,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::WorkerUtilityCredit {
-            endpoint,
-            event_type,
-            credit_points_delta,
-            reason,
-            external_ref,
-            submission_ids,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_worker_utility_credit(TraceCommonsWorkerUtilityCreditOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                event_type,
-                credit_points_delta,
-                reason,
-                external_ref,
-                submission_ids,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ProcessEvaluationSubmit(args) => {
-            trace_commons_process_evaluation_submit(TraceCommonsProcessEvaluationSubmitOptions {
-                endpoint: &args.endpoint,
-                bearer_token_env: &args.bearer_token_env,
-                submission_id: args.submission_id,
-                reason: args.reason,
-                evaluator_name: args.evaluator_name,
-                evaluator_version: args.evaluator_version,
-                labels: args.labels,
-                tool_selection: args.tool_selection,
-                tool_argument_quality: args.tool_argument_quality,
-                tool_ordering: args.tool_ordering,
-                verification: args.verification,
-                side_effect_safety: args.side_effect_safety,
-                overall_score: args.overall_score,
-                utility_credit_points_delta: args.utility_credit_points_delta,
-                utility_external_ref: args.utility_external_ref,
-                json: args.json,
-            })
-            .await
-        }
-        TracesSubcommand::AnalyticsSummary {
-            endpoint,
-            bearer_token_env,
-            json,
-        } => trace_commons_analytics_summary(&endpoint, &bearer_token_env, json).await,
-        TracesSubcommand::MaintenanceRun {
-            endpoint,
-            purpose,
-            dry_run,
-            backfill_db_mirror,
-            index_vectors,
-            purge_expired_before,
-            max_export_age_hours,
-            skip_export_cache_prune,
-            reconcile_db_mirror,
-            verify_audit_chain,
-            bearer_token_env,
-            json,
-        } => {
-            let options = TraceCommonsMaintenanceOptions {
-                purpose,
-                dry_run,
-                backfill_db_mirror,
-                index_vectors,
-                reconcile_db_mirror,
-                verify_audit_chain,
-                prune_export_cache: !skip_export_cache_prune,
-                max_export_age_hours,
-                purge_expired_before,
-            };
-            trace_commons_maintenance_run(&endpoint, &bearer_token_env, options, json).await
-        }
-        TracesSubcommand::WorkerRetentionMaintenance {
-            endpoint,
-            purpose,
-            dry_run,
-            purge_expired_before,
-            max_export_age_hours,
-            skip_export_cache_prune,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_retention_maintenance_run(TraceCommonsRetentionMaintenanceOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                dry_run,
-                prune_export_cache: !skip_export_cache_prune,
-                max_export_age_hours,
-                purge_expired_before,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::RetentionJobsList {
-            endpoint,
-            limit,
-            status,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_retention_jobs_list(&endpoint, &bearer_token_env, limit, status, json)
-                .await
-        }
-        TracesSubcommand::RetentionJobItems {
-            endpoint,
-            retention_job_id,
-            limit,
-            action,
-            status,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_retention_job_items(
-                &endpoint,
-                &bearer_token_env,
-                retention_job_id,
-                limit,
-                action,
-                status,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::ExportAccessGrantsList {
-            endpoint,
-            limit,
-            status,
-            dataset_kind,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_export_access_grants_list(
-                &endpoint,
-                &bearer_token_env,
-                limit,
-                status,
-                dataset_kind,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::ExportJobsList {
-            endpoint,
-            limit,
-            status,
-            dataset_kind,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_export_jobs_list(
-                &endpoint,
-                &bearer_token_env,
-                limit,
-                status,
-                dataset_kind,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::WorkerVectorIndex {
-            endpoint,
-            purpose,
-            dry_run,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_vector_index_run(&endpoint, &bearer_token_env, purpose, dry_run, json)
-                .await
-        }
-        TracesSubcommand::BenchmarkConvert {
-            endpoint,
-            purpose,
-            limit,
-            consent_scope,
-            status,
-            privacy_risk,
-            external_ref,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_benchmark_convert(TraceCommonsBenchmarkConvertOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                limit,
-                consent_scope,
-                status,
-                privacy_risk,
-                external_ref,
-                json,
-                path: "/v1/benchmarks/convert",
-            })
-            .await
-        }
-        TracesSubcommand::WorkerBenchmarkConvert {
-            endpoint,
-            purpose,
-            limit,
-            consent_scope,
-            status,
-            privacy_risk,
-            external_ref,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_benchmark_convert(TraceCommonsBenchmarkConvertOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                limit,
-                consent_scope,
-                status,
-                privacy_risk,
-                external_ref,
-                json,
-                path: "/v1/workers/benchmark-convert",
-            })
-            .await
-        }
-        TracesSubcommand::BenchmarkLifecycleUpdate {
-            endpoint,
-            conversion_id,
-            registry_status,
-            registry_ref,
-            published_at,
-            evaluation_status,
-            evaluator_ref,
-            evaluated_at,
-            score,
-            pass_count,
-            fail_count,
-            reason,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_benchmark_lifecycle_update(TraceCommonsBenchmarkLifecycleUpdateOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                conversion_id,
-                registry_status,
-                registry_ref,
-                published_at,
-                evaluation_status,
-                evaluator_ref,
-                evaluated_at,
-                score,
-                pass_count,
-                fail_count,
-                reason,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::ReplayDatasetExport {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_replay_dataset_export(TraceCommonsReplayDatasetExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                method: Method::GET,
-                path: "/v1/datasets/replay",
-            })
-            .await
-        }
-        TracesSubcommand::WorkerReplayDatasetExport {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_replay_dataset_export(TraceCommonsReplayDatasetExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                method: Method::GET,
-                path: "/v1/workers/replay-export",
-            })
-            .await
-        }
-        TracesSubcommand::ReplayExportManifests {
-            endpoint,
-            bearer_token_env,
-            json,
-        } => trace_commons_replay_export_manifests(&endpoint, &bearer_token_env, json).await,
-        TracesSubcommand::ConfigStatus {
-            endpoint,
-            bearer_token_env,
-        } => trace_commons_config_status(&endpoint, &bearer_token_env).await,
-        TracesSubcommand::OperationalSummary {
-            endpoint,
-            bearer_token_env,
-            json,
-        } => trace_commons_operational_summary(&endpoint, &bearer_token_env, json).await,
-        TracesSubcommand::TenantPolicyGet {
-            endpoint,
-            bearer_token_env,
-            json,
-        } => trace_commons_tenant_policy_get(&endpoint, &bearer_token_env, json).await,
-        TracesSubcommand::TenantPolicySet {
-            endpoint,
-            policy_version,
-            allowed_consent_scopes,
-            allowed_uses,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_tenant_policy_set(
-                &endpoint,
-                &bearer_token_env,
-                policy_version,
-                allowed_consent_scopes,
-                allowed_uses,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::TenantAccessGrantsList {
-            endpoint,
-            limit,
-            status,
-            role,
-            principal_ref,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_tenant_access_grants_list(
-                &endpoint,
-                &bearer_token_env,
-                limit,
-                status,
-                role,
-                principal_ref,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::TenantPrincipalRef {
-            token_env,
-            signed_tenant_id,
-            signed_actor_ref,
-            json,
-        } => trace_commons_tenant_principal_ref(
-            token_env.as_deref(),
-            signed_tenant_id.as_deref(),
-            signed_actor_ref.as_deref(),
-            json,
-        ),
-        TracesSubcommand::TenantAccessGrantCreate {
-            endpoint,
-            principal_ref,
-            role,
-            grant_id,
-            allowed_consent_scopes,
-            allowed_uses,
-            issuer,
-            audience,
-            subject,
-            issued_at,
-            expires_at,
-            reason,
-            metadata,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_tenant_access_grant_create(TraceCommonsTenantAccessGrantCreateOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                principal_ref,
-                role,
-                grant_id,
-                allowed_consent_scopes,
-                allowed_uses,
-                issuer,
-                audience,
-                subject,
-                issued_at,
-                expires_at,
-                reason,
-                metadata,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::TenantAccessGrantRevoke {
-            endpoint,
-            grant_id,
-            reason,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_tenant_access_grant_revoke(
-                &endpoint,
-                &bearer_token_env,
-                grant_id,
-                reason,
-                json,
-            )
-            .await
-        }
-        TracesSubcommand::RankerTrainingCandidates {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_ranker_training_export(TraceCommonsRankerTrainingExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                path: "/v1/ranker/training-candidates",
-                output_label: "ranker training candidates",
-                item_field: "candidates",
-            })
-            .await
-        }
-        TracesSubcommand::WorkerRankerTrainingCandidates {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_ranker_training_export(TraceCommonsRankerTrainingExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                path: "/v1/workers/ranker/training-candidates",
-                output_label: "ranker training candidates",
-                item_field: "candidates",
-            })
-            .await
-        }
-        TracesSubcommand::RankerTrainingPairs {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_ranker_training_export(TraceCommonsRankerTrainingExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                path: "/v1/ranker/training-pairs",
-                output_label: "ranker training pairs",
-                item_field: "pairs",
-            })
-            .await
-        }
-        TracesSubcommand::WorkerRankerTrainingPairs {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            privacy_risk,
-            limit,
-            output,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_ranker_training_export(TraceCommonsRankerTrainingExportOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                privacy_risk,
-                limit,
-                output,
-                json,
-                path: "/v1/workers/ranker/training-pairs",
-                output_label: "ranker training pairs",
-                item_field: "pairs",
-            })
-            .await
-        }
-        TracesSubcommand::AuditEvents {
-            endpoint,
-            limit,
-            bearer_token_env,
-            json,
-        } => trace_commons_audit_events(&endpoint, &bearer_token_env, limit, json).await,
-        TracesSubcommand::ListTraces {
-            endpoint,
-            purpose,
-            consent_scope,
-            status,
-            limit,
-            coverage_tag,
-            tool,
-            privacy_risk,
-            bearer_token_env,
-            json,
-        } => {
-            trace_commons_list_traces(TraceCommonsListTracesOptions {
-                endpoint: &endpoint,
-                bearer_token_env: &bearer_token_env,
-                purpose,
-                consent_scope,
-                status,
-                limit,
-                coverage_tag,
-                tool,
-                privacy_risk,
-                json,
-            })
-            .await
-        }
-        TracesSubcommand::PrivacyFilterCanary {
-            text,
-            timeout_seconds,
-            json,
-        } => privacy_filter_canary(&text, timeout_seconds, json).await,
+        v @ TracesSubcommand::OptIn { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::OptOut { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Status { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Preview { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Enqueue { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::FlushQueue { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::QueueStatus { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Credit { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Submit { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::ListSubmissions { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::Revoke { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::IngestHealth { .. } => contributor::dispatch(v).await,
+        v @ TracesSubcommand::QuarantineList { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ActiveLearningReviewQueue { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ReviewDecision { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ReviewLeaseClaim { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ReviewLeaseClaimNext { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ReviewLeaseClaimBatch { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::ReviewLeaseRelease { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::AppendCreditEvent { .. } => reviewer::dispatch(v).await,
+        v @ TracesSubcommand::WorkerUtilityCredit { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerRetentionMaintenance { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerVectorIndex { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerBenchmarkConvert { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerReplayDatasetExport { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerRankerTrainingCandidates { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::WorkerRankerTrainingPairs { .. } => worker::dispatch(v).await,
+        v @ TracesSubcommand::ProcessEvaluationSubmit(_) => worker::dispatch(v).await,
+        v @ TracesSubcommand::MaintenanceRun { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::RetentionJobsList { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::RetentionJobItems { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::ExportAccessGrantsList { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::ExportJobsList { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::BenchmarkConvert { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::BenchmarkLifecycleUpdate { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::ReplayDatasetExport { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::ReplayExportManifests { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::AnalyticsSummary { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::OperationalSummary { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::ConfigStatus { .. } => admin::dispatch(v).await,
+        v @ TracesSubcommand::TenantPolicyGet { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::TenantPolicySet { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::TenantAccessGrantsList { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::TenantPrincipalRef { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::TenantAccessGrantCreate { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::TenantAccessGrantRevoke { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::RankerTrainingCandidates { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::RankerTrainingPairs { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::AuditEvents { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::ListTraces { .. } => tenant::dispatch(v).await,
+        v @ TracesSubcommand::PrivacyFilterCanary { .. } => tenant::dispatch(v).await,
     }
 }
